requirements.txt
amqp==5.1.1
asgiref==3.5.0
astroid==2.9.3
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.6
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cloudinary-storage==0.3.0
django-extensions==3.2.1
django-jazzmin==2.4.9
django-timezone-field==4.2.3
gunicorn==20.1.0
idna==3.4
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.1
redis==4.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.13.3
